Generic_Attributes: Update all device's __init__ method to be decoraited by @set_passed_properties() in most cases I used @set_passed_properties(keep_names = []) indicating that none of the kw parameters were to be entered into the properties field.
In the case of the Camera device, all of the information that is currently stored
as attributes to the device is also being stored in this field (since this is
really where it belongs).  The NIST imaging software will pull all camera info
from this connection table field rather than the device attributes.  The device
attributes will currently be left in place for backwards compatability.