feature-CamAcqROI: Added acquisition_ROI
I added a new device property to the Camera class to allow setting and
tracking of the read out sensor area on a per shot basis.
There is a new method, set_acquisition_ROI, that sets the new propery
while sweeping some of set_property's features under the rug.

DSB