NI_DAQmx: hacking on wait monitor. Not yet functional, some issues to iron out with incomplete sample detection and small buffers.
Plan is to

* acquire pulse width instead of semiperiod, this will entail
only one read event per pulse rather than two, solving the issue of data
coming faster than it can be read and filling the buffer. Or maybe we will
still do semiperiod, but will read 2 samples at a time as David Meyer has had
sucess with. The advantage of this latter approach is that we do not need to
store the pulse duration in the h5 file in order to take it into account -
we can measure it.

* Determine whether a device has incomplete sample detection empirically.
  In get_capabilities.py we can create a task that outputs on a DO and reads in
  a counter, and we can see what we get and store whether it looks like
  incomplete sample detection. This should be achievable without asking the user
  to connect back any cables physically, as we should be able to configure
  them to be connected internally. This will not work if there are any devices
  out there with counters but no digital outs!