novatech-default-baudrate: Use David Meyer's technique of iterating over all baud rates at startup if communication cannot be established at the current baud rate.
This resolves an issue whereby restarting a tab in BLACS does not cleanly shut
it down, and so the Novatech's baud rate would not be reset to te default value
upon restart. The startup code would then not be able to communicate with it.

BLACS should try to cleanly shut down tabs upon restart, if possible, but
that is a separate issue.