pulse-width-fixes: Resolve issue #33.
Pulseblasters now have a long delay time computed from their core clock freq,
which has been added as an instance attribute to every subclass.
(this is messy, as it is duplicated in the BLACS worker classes too. It should be
passed as a connection table property ideally, but this will have to wait for a
more general 'pulseblaster unification' change I think, along the lines of the
NI DAQ unification).

For each clock tick, the high time is half the clock period if
`pulse_width='symmetric'` (the default), the minimum possible if
`pulse_width='minimum'`, or a fixed value if `pulse_width` is given as a number.

The high time is then clipped to self.long_delay if it is larger, in which case
the pulse_width argument is not honoured, but 57 second high times ought to be
visible on scopes and not too fast for any devices, so this is preferable to
the alternative of adding LONG_DELAY instructions to *both* the high and low
times of the clock ticks.

The low time is then computed as whatever is left, and split into a LONG_DELAY
instruction plus the remainder as an END_LOOP instruction, as before.

If the PulseBlaster has no external clocks for this clock tick, it uses a
LONG_DELAY plus the remainder as a CONTINUE instruction, as before.