bugfix: Fix a deadlock restarting IMAQdxCamera.
If restart was clicked whilst continuously acquiring, the ImageReceiver could
be waiting on the previous frame to be updated, and thus would not respond to
the request to shutdown. However since shutdown() was called from the main thread,
it would be waiting forever.

Now it regularly checks self.stopping (an attribute set bt the parent class
zprocess.ZMQServer when a shutdown has been requested) while waiting for the
update event to process, so that it does not deadlock.

Also, if the update event runs but finds that it has already been shut down,
it does nothing, since the Qt widgets may have been destroyed already.