novatech_phase_mode_fix: Fix issue #37, incorrect transition to manual/buffered semantics for novatech.
It appears that it was a misunderstanding to think of 'm 0' 'm a' and 'm n' as
three options for the phase mode (with 'm 0' meaning 'default' - corresponding
to one of the others). Rather, 'm n' and 'm a' are the two options
for the phase mode, and 'm 0' and 'm t' are for switching between table mode
and single-tone mode.

Therefore this PR calls 'm 0' at startup, and then one of 'm a' or 'm n' to set
the phase mode, which is now by default 'continuous', with 'default' having been
removed as an option, but being interpreted as 'continuous' for the case of
already-compiled shots that used it.

Thereafter, transition between single-tone mode and table mode is done with
'm 0' and 'm t', with no further calls to set the phase mode.

Also added a better explanation of the 'double clutch' hack.