NI_DAQmx: Write all DO at once instead of iterating over ports.
This resolves an issue raised in the mailing list:
https://groups.google.com/forum/#!topic/labscriptsuite/pO6CGa4GsRI

in which some devices are unhappy about the explicit buffer configuration
required when making multiple writes of data to the same task. By making a
single write only, we allow the DAQmx driver to make decisions about buffers
itself, which the NI-PCI-DIO-32HS seems to require.