AndorCamera: In andor_solis.py, added two SDK methods for downloading oldest 16 bit frames in the buffer. Also added several status code errors that invalidate operations. A third SDK method inspects the max Images that can be trasnferred on a DMA transaction was also added.
In andor utils.py, merged a lot of the logic coming out of tests in the lab
mostly relevant to crop mode download acquisition and fast kinetics modes.
Incorporated changes to temperature control logic from R.P. Anderson and a few
things to the EMCCD settings. There are a few changes in the vertical shifting
speed setup as well to differentiate FK from other modes. Simplified some of
the readout setup lines. All of this needs testing!

In blacks_workers.py, added some printing statements as well as the FK mode
download lines. These handle the correct shaping as the data comes.

runviewer_parsers.py was still the same as when I copied it from another device