Add NI-DAQmx documentation.
This employs apidoc to auto-generate a custom templated summary of all
the auto-generated sub-class models. It modifies conf.py to auto-run
apidoc to create the template. Therefore new sub-classes that are added
will automatically have the appropriate documentation generated.

Note: apidoc creates a large number of cruft docfiles for all modules in labscript_devices using this specific custom template. This is because I
haven't figured out how to get the autodoc directives to work without
full specification of the module from the top level. This could be worked
around by the apidoc exclude option, but it only uses very basic fnmatch
syntax that can't seem to be fudged to only select a single submodule.