Decreasing programming time of PB devices
When programming large instruction sets the speed of the loops in transition_to_buffered starts to bog down and results in long initial loading times at the beginning of a brand new shot. Since large for loops are slow the for loop was swapped out with map() instead. This results in almost immediate loading times and shaves seconds off of starting up a different experiment.