Modified NI_DAQmx to handle differentially terminated analog inputs.
Behavior is to either have all the inputs as single-ended or differential.
Also modified `get_capabilities.py` a little more so that it will correctly
set the default AI termination for devices that do not support RSE.

Tested with a NI USB 6366. Was able to simultaneously sample at 2MS/s
on two channels (the max spec for this device).