Added a minimum_clock_high_time attribute the the NovaTechDDS9M labscript device class.
This is the corresponding change for labscript-suite/labscript#88

Technically I think this can be as low as 100ns but I'm concerned that some of the tri-state buffer hardware (or similar) that some labs have in front of their NovaTechs might mean this is actually slower. 1us seems like a safe bet while allowing people to command output on other clocklines within a reasonable time frame.