Changed SpinnakerCamera StreamBufferCountMode
Previously, the buffer count mode for the Spinnaker Camera was set to Auto. This feature has been depreciated so I changed it to be manual, and manually set the buffer count. In the case where bufferCount == 1, I set it to 3 because the Spinnaker SDK mentioned that 3 was the minimum buffer count.